+++ packages/core/test/feedback/FeedbackWidgetManager.test.tsx
@@ -4,9 +4,9 @@ import * as React from 'react';
 import { Text } from 'react-native';
 
 import { defaultConfiguration } from '../../src/js/feedback/defaults';
-import { FeedbackWidgetProvider, resetFeedbackWidgetManager, showFeedbackWidget } from '../../src/js/feedback/FeedbackWidgetManager';
+import { FeedbackWidgetProvider, hideFeedbackButton,resetFeedbackButtonManager, resetFeedbackWidgetManager, showFeedbackButton, showFeedbackWidget } from '../../src/js/feedback/FeedbackWidgetManager';
 import { feedbackIntegration } from '../../src/js/feedback/integration';
-import { AUTO_INJECT_FEEDBACK_INTEGRATION_NAME } from '../../src/js/feedback/lazy';
+import { AUTO_INJECT_FEEDBACK_BUTTON_INTEGRATION_NAME,AUTO_INJECT_FEEDBACK_INTEGRATION_NAME } from '../../src/js/feedback/lazy';
 import { isModalSupported } from '../../src/js/feedback/utils';
 import { getDefaultTestClientOptions, TestClient } from '../mocks/client';
 
@@ -114,7 +114,7 @@ describe('FeedbackWidgetManager', () => {
 
     showFeedbackWidget();
 
-    expect(consoleWarnSpy).toHaveBeenLastCalledWith('[Sentry] FeedbackWidget requires `Sentry.wrap(RootComponent)` to be called before `showFeedbackWidget()`.');
+    expect(consoleWarnSpy).toHaveBeenLastCalledWith(`[Sentry] FeedbackWidget requires 'Sentry.wrap(RootComponent)' to be called before 'showFeedbackWidget()'.`);
   });
 
   it('showFeedbackWidget does not warn about missing feedback provider when FeedbackWidgetProvider is used', () => {
@@ -130,12 +130,78 @@ describe('FeedbackWidgetManager', () => {
 
     expect(consoleWarnSpy).not.toHaveBeenCalled();
   });
-  
+
   it('showFeedbackWidget adds the feedbackIntegration to the client', () => {
     mockedIsModalSupported.mockReturnValue(true);
 
     showFeedbackWidget();
 
     expect(getClient().getIntegrationByName(AUTO_INJECT_FEEDBACK_INTEGRATION_NAME)).toBeDefined();
+  });
+});
+
+describe('FeedbackButtonManager', () => {
+
+  beforeEach(() => {
+    const client = new TestClient(getDefaultTestClientOptions());
+    setCurrentClient(client);
+    client.init();
+    consoleWarnSpy.mockReset();
+    resetFeedbackButtonManager();
+  });
+
+  it('showFeedbackButton displays the button when FeedbackWidgetProvider is used', () => {
+    const { getByText } = render(
+      <FeedbackWidgetProvider>
+        <Text>App Components</Text>
+      </FeedbackWidgetProvider>
+    );
+
+    showFeedbackButton();
+
+    expect(getByText('Report a Bug')).toBeTruthy();
+  });
+
+  it('hideFeedbackButton hides the button', () => {
+    const { queryByText } = render(
+      <FeedbackWidgetProvider>
+        <Text>App Components</Text>
+      </FeedbackWidgetProvider>
+    );
+
+    showFeedbackButton();
+    hideFeedbackButton();
+
+    expect(queryByText('Report a Bug')).toBeNull();
+  });
+
+  it('showFeedbackButton does not throw an error when FeedbackWidgetProvider is not used', () => {
+    expect(() => {
+      showFeedbackButton();
+    }).not.toThrow();
+  });
+
+  it('showFeedbackButton warns about missing feedback provider', () => {
+    showFeedbackButton();
+
+    expect(consoleWarnSpy).toHaveBeenLastCalledWith(`[Sentry] FeedbackButton requires 'Sentry.wrap(RootComponent)' to be called before 'showFeedbackButton()'.`);
+  });
+
+  it('showFeedbackButton does not warn about missing feedback provider when FeedbackWidgetProvider is used', () => {
+    render(
+      <FeedbackWidgetProvider>
+        <Text>App Components</Text>
+      </FeedbackWidgetProvider>
+    );
+
+    showFeedbackButton();
+
+    expect(consoleWarnSpy).not.toHaveBeenCalled();
+  });
+
+  it('showFeedbackButton adds the feedbackIntegration to the client', () => {
+    showFeedbackButton();
+
+    expect(getClient().getIntegrationByName(AUTO_INJECT_FEEDBACK_BUTTON_INTEGRATION_NAME)).toBeDefined();
   });
 });