+++ samples/expo/app/(tabs)/index.tsx
@@ -70,6 +70,12 @@ export default function TabOneScreen() {
         }}
       />
       <Button
+        title="Show feedback button"
+        onPress={() => {
+          Sentry.showFeedbackButton();
+        }}
+      />
+      <Button
         title="Set Scope Properties"
         onPress={() => {
           setScopeProperties();