+++ samples/react-native/src/Screens/ErrorsScreen.tsx
@@ -31,6 +31,7 @@ const ErrorsScreen = (_props: Props) => {
   // Show bad code inside error boundary to trigger it.
   const [showBadCode, setShowBadCode] = React.useState(false);
   const [isFeedbackVisible, setFeedbackVisible] = React.useState(false);
+  const [isFeedbackButtonVisible, setFeedbackButtonVisible] = React.useState(false);
 
   const errorBoundaryFallback: FallbackRender = ({ eventId }) => (
     <Text>Error boundary caught with event id: {eventId}</Text>
@@ -223,6 +224,18 @@ const ErrorsScreen = (_props: Props) => {
           title="Feedback form (auto)"
           onPress={() => {
             Sentry.showFeedbackWidget();
+          }}
+        />
+        <Button
+          title="Show/Hide Feedback Button"
+          onPress={() => {
+            if (isFeedbackButtonVisible) {
+              Sentry.hideFeedbackButton();
+              setFeedbackButtonVisible(false);
+            } else {
+              Sentry.showFeedbackButton();
+              setFeedbackButtonVisible(true);
+            }
           }}
         />
         <Button